Add OrganizeCharts Json View Model & Controller

+++ AIKI.CO.HelpDesk.WebAPI/Controllers/BaseApiController.cs
@@ -21,14 +21,16 @@ namespace AIKI.CO.HelpDesk.WebAPI.Controllers
         protected readonly AppSettings _appSettings;
         protected readonly IMapper _map;
         protected readonly IService<T, V> _service;
+        protected bool _isReadOnly;
         public BaseCRUDApiController(
             IMapper map,
             IOptions<AppSettings> appSettings,
-            IService<T, V> service)
+            IService<T, V> service, bool isReadOnly = false)
         {
             _map = map;
             _appSettings = appSettings.Value;
             _service = service;
+            _isReadOnly = isReadOnly;
         }
 
         [HttpGet]
@@ -56,6 +58,7 @@ namespace AIKI.CO.HelpDesk.WebAPI.Controllers
         [Produces("application/json")]
         public virtual async Task<IActionResult> Post([FromBody] V request)
         {
+            if (_isReadOnly) return BadRequest("Entity is ReadOnly");
             if (!ModelState.IsValid) return BadRequest(ModelState);
             var result = await _service.AddRecord(request);
             if (result > 0)
@@ -66,6 +69,7 @@ namespace AIKI.CO.HelpDesk.WebAPI.Controllers
         [HttpPut]
         public virtual async Task<IActionResult> Put([FromBody] V request)
         {
+            if (_isReadOnly) return BadRequest("Entity is ReadOnly");
             if (!ModelState.IsValid) return BadRequest(ModelState);
             var existsRecord = await _service.GetById(request.id);
             if (existsRecord == null) return NotFound();
@@ -77,6 +81,7 @@ namespace AIKI.CO.HelpDesk.WebAPI.Controllers
         [HttpPatch("{id:guid}")]
         public virtual async Task<IActionResult> Patch([FromRoute] Guid id, [FromBody] JsonPatchDocument<V> patchDoc)
         {
+            if (_isReadOnly) return BadRequest("Entity is ReadOnly");
             if (patchDoc == null) return BadRequest();
             var founded = await _service.GetById(id);
             if (founded == null) return NotFound();
@@ -95,6 +100,7 @@ namespace AIKI.CO.HelpDesk.WebAPI.Controllers
         [HttpDelete("{id:guid}")]
         public virtual async Task<IActionResult> Delete([FromRoute] Guid id)
         {
+            if (_isReadOnly) return BadRequest("Entity is ReadOnly");
             var result = await _service.DeleteRecord(id);
             if (result > 0)
                 return Ok(id);
